% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_api.R
\name{nppes_api}
\alias{nppes_api}
\alias{print.nppes_api}
\title{NPPES API}
\usage{
nppes_api(query)

\method{print}{nppes_api}(x, ...)
}
\arguments{
\item{query}{Character string representing a valid query.}
}
\value{
An bject of class \code{"nppes_api"}; essentially, an R list with 
various components depending on the submitted query.
}
\description{
Query the NPPES NPI registry.
}
\examples{
nppes_api("city=dayton")
nppes_api("number=1124021324")
}
