% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppes_api.R
\name{nppes_fields}
\alias{nppes_fields}
\title{NPPES Fields}
\usage{
nppes_fields()
}
\description{
Print available NPPES field values that can be queried.
}
\references{
A vector containing the names of the available for querying.
}
